requirements.txt
tensorflow>=2.13.0
opencv-python>=4.8.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
lime>=0.2.0
jupyter>=1.0.0
